bench: benchmark-model subcommand + commit the RecurrentGemma manifest
models/ is no longer gitignored: the manifest is 114 KB of spec +
safetensors descriptors, so committing it makes the converted model
runnable on any machine that has the checkpoint at
gemma/recurrentgemma-2b (relative descriptor paths resolve against
the manifest's directory; regenerate with --checkpoint for other
layouts). The converter now stamps the manifest fp32 -- the
published checkpoint's native dtype (the earlier bf16 was a wrong
assumption); loaders can still override programmatically.

texmo.py benchmark-model: greedy generation of a stored model at
several lengths (default 128,256). Each run does identical fixed
work (fresh states + prompt) plus N tokens with a warm jit, so
(T[2N]-T[N])/N is the steady-state per-token rate and 2*T[N]-T[2N]
the per-run overhead; load and jit-compile times are reported
separately, plus system name and device for cross-machine
comparison. Runs at the manifest's precision -- deterministic and
comparable by construction. Caveat learned on the first run: the
two-point split assumes a stationary rate -- run it on an otherwise
idle machine.

+++ texmo/convert/recurrentgemma.py
@@ -247,7 +247,9 @@ def main():
     mapper = _Mapper(args.checkpoint, out_dir, weight_map)
     doc = {
         'spec': spec,
-        'precision': 'bf16',
+        # The published checkpoint is fp32 (config torch_dtype);
+        # loaders can override (e.g. Precision.BF16 to halve RAM).
+        'precision': 'fp32',
         'weights': build_weights(cfg, mapper),
     }
